refactor: JWT 관련 클래스 주석 수정 및 예외 클래스 패키지 변경
