--- src/main/java/com/project/food_ordering_service/global/utils/jwt/JwtExceptionHandlerFilter.java
@@ -1,4 +1,0 @@
-package com.project.food_ordering_service.global.utils.jwt;
-
-public class JwtExceptionHandlerFilter {
-}
