+++ src/main/java/com/project/food_ordering_service/global/utils/jwt/exception/InvalidTokenException.java
@@ -1,4 +1,4 @@
-package com.project.food_ordering_service.global.utils.jwt;
+package com.project.food_ordering_service.global.utils.jwt.exception;
 
 public class InvalidTokenException extends RuntimeException {
     public InvalidTokenException() {
